requirements.txt
Faker==18.9.0
python-dotenv==1.0.0
requests==2.28.2
